rewards headache

+++ src/main/java/org/launchcode/professionalprocrastinators/models/data/RewardsRepository.java
@@ -1,7 +1,9 @@
 package org.launchcode.professionalprocrastinators.models.data;
 
+// RewardsRepository.java
 import org.launchcode.professionalprocrastinators.models.*;
-import org.springframework.data.repository.*;
+import org.springframework.data.repository.CrudRepository;
 
-public interface RewardsRepository extends CrudRepository<Rewards, Long> {
+public interface RewardsRepository extends CrudRepository<Rewards, Integer> {
+
 }
